chore: add try catch
The change queue is failing because a project has been closed already. We can add `lix.settled()` APIs in tests but that will be a never ending katz und maus spiel.

The solution is to re-work the change queue, don't depent on an app writing files, and getting rid of `lix.settled()`.

https://github.com/opral/lix-sdk/issues/35